(refactor) 3단계 - 각 클래스 별 테스트 작성, 검증 클래스를 도메인에 작성, 람다식으로 변경
